% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_enrichment.r
\name{allele_enrichment}
\alias{allele_enrichment}
\title{Calculates enrichment of desirable allele. Given genotypes (gt),
a vector (w) of weight values (equal length to the 
number of individuals, a vector (v) of values describing
the importance of each locus, estimates an index of enrichment of
preferred alleles. Assumes genotype fitness 2 > 1 > 0.}
\usage{
allele_enrichment(gt, w = NULL, loc = NULL, rec = FALSE)
}
\arguments{
\item{gt}{- genotype matrix for adaptive loci (fitness 2 > 1 > 0)}

\item{w}{- vector of individual weights}

\item{loc}{- vector of locus adaptation values}

\item{rec}{- if true, fitness 2 > 1 = 0}
}
\value{
An index of adaptedness, float value
}
\description{
Calculates enrichment of desirable allele. Given genotypes (gt),
a vector (w) of weight values (equal length to the 
number of individuals, a vector (v) of values describing
the importance of each locus, estimates an index of enrichment of
preferred alleles. Assumes genotype fitness 2 > 1 > 0.
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

