% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_nondominated_archives.r
\name{combine_nondominated_archives}
\alias{combine_nondominated_archives}
\title{Combines two non-dominated archives, output from 
multiobjective optimization function}
\usage{
combine_nondominated_archives(a1, a2)
}
\arguments{
\item{a1}{- a non-dominated archive}

\item{a2}{- a non-dominated archive}
}
\value{
a list of output
}
\description{
Combines two non-dominated archives, output from 
multiobjective optimization function
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

