% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_measure_value.r
\name{generate_measure_value}
\alias{generate_measure_value}
\title{Wraps functions that are used to calculate summaries}
\usage{
generate_measure_value(v, measure = NULL, w = NULL, q = NULL,
  loc = NULL, disp = 0)
}
\arguments{
\item{v}{- gt or sm}

\item{measure}{- "nei" or "shannon" or "mean_similarity"}

\item{w}{- weights}

\item{q}{- q value for shannon diversity}
}
\value{
summary value (if similarity, summary value * -1)
}
\description{
Wraps functions that are used to calculate summaries
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

