% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_matrix_by_weights.r
\name{genotype_matrix_by_weights}
\alias{genotype_matrix_by_weights}
\title{Transforms a genotype matrix by a vector of weights}
\usage{
genotype_matrix_by_weights(gt, w)
}
\arguments{
\item{gt}{- genotype matrix (samples by loci)}

\item{w}{- weights}
}
\value{
gt matrix transformed according to weights
}
\description{
Transforms a genotype matrix by a vector of weights
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

