% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_accept_reject.r
\name{multi_accept_reject}
\alias{multi_accept_reject}
\title{Determines acceptance or rejection of a proposal}
\usage{
multi_accept_reject(summary_1 = summary_1, summary_2 = summary_2,
  proposal_summary_1 = proposal_summary_1,
  proposal_summary_2 = proposal_summary_2, t, p_depends_delta = FALSE,
  c1 = 1, c2 = 1, cboth = 1)
}
\arguments{
\item{t}{- current temperature}

\item{p_depends_delta}{- if TRUE, accept depends on difference between new and old}

\item{c1}{- multiplier for numerator when proposal for value 1 worse, 2 better}

\item{c2}{- multiplier for numerator when proposal for value 2 worse, 1 better}

\item{cboth}{- multiplier for numerator when proposals for both values are worse}

\item{summary}{- summary of current}

\item{proposal_summary}{- summary of new proposal}
}
\value{
Boolean, true if proposal accepted
}
\description{
Determines acceptance or rejection of a proposal
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

