% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nei_diversity.r
\name{nei_diversity}
\alias{nei_diversity}
\title{Calculates Nei diversity, given a set of genotypes (gt)
and a vector of weight values (equal length to the 
number of individuals}
\usage{
nei_diversity(gt, w = NULL)
}
\arguments{
\item{gt}{- genotype matrix (individuals, loci)}

\item{w}{- vector of weights}
}
\value{
Nei diversity, a float value
}
\description{
Calculates Nei diversity, given a set of genotypes (gt)
and a vector of weight values (equal length to the 
number of individuals
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

