% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_matrix_by_weights.r
\name{pairwise_matrix_by_weights}
\alias{pairwise_matrix_by_weights}
\title{Transforms a pairwise matrix by a vector of weights}
\usage{
pairwise_matrix_by_weights(sm, w)
}
\arguments{
\item{sm}{- pairwise sample matrix (samples by samples)}

\item{w}{- weights}
}
\value{
sm matrix transformed according to weights
}
\description{
Transforms a pairwise matrix by a vector of weights
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

