% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposal_accept_reject.r
\name{proposal_accept_reject}
\alias{proposal_accept_reject}
\title{Determines acceptance or rejection of a proposal}
\usage{
proposal_accept_reject(summary = summary,
  proposal_summary = proposal_summary, t, p_depends_delta = FALSE, c = 1)
}
\arguments{
\item{summary}{- summary of current}

\item{proposal_summary}{- summary of new proposal}

\item{t}{- current temperature}

\item{p_depends_delta}{- if TRUE, accept depends on difference between new and old}

\item{c}{- constant multiplier for numerator (default, c=1)}
}
\value{
Boolean, true if proposal accepted
}
\description{
Determines acceptance or rejection of a proposal
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

