% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propose_initial_weights.r
\name{propose_initial_weights}
\alias{propose_initial_weights}
\title{Proposes initial weights given number of genotypes (length of weights) 
and number of individuals in target population (sum of weights).}
\usage{
propose_initial_weights(N_g, N_t, w_max = NULL, verbose = FALSE)
}
\arguments{
\item{N_g}{- number of individuals with genotypes}

\item{N_t}{- number of individuals in target population}

\item{w_max}{- a vector of maxima for each individual}
}
\value{
An initial, random, set of weights
}
\description{
Proposes initial weights given number of genotypes (length of weights) 
and number of individuals in target population (sum of weights).
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

