% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propose_new_weights.r
\name{propose_new_weights}
\alias{propose_new_weights}
\title{Proposes new weights in the neighborhood of current 
state. Uses current state, constraints (maxima).}
\usage{
propose_new_weights(w, w_max = NULL)
}
\arguments{
\item{w}{- a vector of weights}

\item{w_max}{- a vector of maxima for each individual}
}
\value{
An updated set of weights
}
\description{
Proposes new weights in the neighborhood of current 
state. Uses current state, constraints (maxima).
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

