% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannon_diversity.r
\name{shannon_diversity}
\alias{shannon_diversity}
\title{Calculates Shannon diversity, given a set of genotypes (gt)
and a vector of weight values (equal length to the 
number of genotyped individuals)}
\usage{
shannon_diversity(gt, w = NULL, q = 1)
}
\arguments{
\item{gt}{- genotype matrix (individuals, loci)}

\item{w}{- vector of weights}

\item{q}{- Shannon level (0, 1, or 2)}
}
\value{
Shannon diversity, a float value
}
\description{
Calculates Shannon diversity, given a set of genotypes (gt)
and a vector of weight values (equal length to the 
number of genotyped individuals)
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

