% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_scheduler.r
\name{temp_scheduler}
\alias{temp_scheduler}
\title{Determines the temperature for annealing}
\usage{
temp_scheduler(s, max_steps, max_t)
}
\arguments{
\item{s}{- current time step}

\item{max_steps}{- maximum step}
}
\value{
temperature, floating point value
}
\description{
Determines the temperature for annealing
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

