% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_mean_of_absolute_difference.r
\name{weighted_mean_of_absolute_difference}
\alias{weighted_mean_of_absolute_difference}
\title{Calculates a mean value for a nominated vector (vs), given
a set of weights (w). Returns the weighted mean of the absolute 
values of differences between vector values and a 'displacement' 
This could be used, e.g., to minimize the mean difference between  
temperature of origin for each sample, and temperate of a site.}
\usage{
weighted_mean_of_absolute_difference(vs, w, disp = 0)
}
\arguments{
\item{vs}{- genotype matrix (individuals, loci)}

\item{w}{- vector of weights}

\item{disp}{- opt value -- displacement from zero}
}
\value{
difference
}
\description{
Calculates a mean value for a nominated vector (vs), given
a set of weights (w). Returns the weighted mean of the absolute 
values of differences between vector values and a 'displacement' 
This could be used, e.g., to minimize the mean difference between  
temperature of origin for each sample, and temperate of a site.
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

