% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_mean_of_pairwise_matrix.r
\name{weighted_mean_of_pairwise_matrix}
\alias{weighted_mean_of_pairwise_matrix}
\title{Calculates the mean value for pairwise matrix, 
when supplied with weights that should be 
applied to the individuals}
\usage{
weighted_mean_of_pairwise_matrix(sm, w = NULL)
}
\arguments{
\item{sm}{- pairwise matrix of values}

\item{w}{- vector of weights}
}
\value{
mean similarity
}
\description{
Calculates the mean value for pairwise matrix, 
when supplied with weights that should be 
applied to the individuals
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

