% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_mean_of_vector.r
\name{weighted_mean_of_vector}
\alias{weighted_mean_of_vector}
\title{Calculates a mean value for a nominated vector (vs), given
a set of weights (w). Returns the absolute value of the 
difference between this weighted mean, and a nominated value
(diff). This could be used to minimize the use of individuals 
low levels of heterozygosity, or to place arbitrary 
constraints on genotype composition}
\usage{
weighted_mean_of_vector(vs, w, disp = 0)
}
\arguments{
\item{vs}{- genotype matrix (individuals, loci)}

\item{w}{- vector of weights}

\item{disp}{- opt value -- displacement from zero}
}
\value{
difference
}
\description{
Calculates a mean value for a nominated vector (vs), given
a set of weights (w). Returns the absolute value of the 
difference between this weighted mean, and a nominated value
(diff). This could be used to minimize the use of individuals 
low levels of heterozygosity, or to place arbitrary 
constraints on genotype composition
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

