% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_multi_objective.r
\name{optimize_multi_objective}
\alias{optimize_multi_objective}
\title{Performs simulated annealing to find mixture of individuals 
that maximize a function (e.g. a measure of genetic diversity)}
\usage{
optimize_multi_objective(v1 = NULL, v2 = NULL, measure_1 = NULL,
  measure_2 = NULL, max_steps = 10000, N_t = NULL,
  initial_weights = NULL, weights_max = NULL, weights_min = NULL,
  max_t = 1, q = NULL, p_depends_delta = FALSE, disp = 0, c1 = 1,
  c2 = 1, cboth = 1, nda = FALSE, min_t = 0, nd_samples = 100)
}
\arguments{
\item{initial_weights}{- a vector of initial weights}

\item{weights_max}{- maximum weights for each individual}

\item{weights_min}{- minimum weights for each individual}

\item{max_t}{- maximum temperature}

\item{gt}{- a genotype matrix (samples, loci)}
}
\value{
a list of output
}
\description{
Performs simulated annealing to find mixture of individuals 
that maximize a function (e.g. a measure of genetic diversity)
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

