% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_single_objective.r
\name{optimize_single_objective}
\alias{optimize_single_objective}
\title{Performs simulated annealing to find mixture of individuals 
that maximize a function (e.g. a measure of genetic diversity)}
\usage{
optimize_single_objective(gt = NULL, sm = NULL, measure = NULL,
  max_steps = 10000, N_t = NULL, initial_weights = NULL,
  weights_max = NULL, max_t = 1, q = NULL, p_depends_delta = FALSE,
  disp = 0)
}
\arguments{
\item{gt}{- a genotype matrix (samples, loci)}

\item{initial_weights}{- a vector of initial weights}

\item{weights_max}{- maximum weights for each individual}

\item{max_t}{- maximum temperature}
}
\value{
a list of output
}
\description{
Performs simulated annealing to find mixture of individuals 
that maximize a function (e.g. a measure of genetic diversity)
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

